+++ math/chineseRemainderTheorem.cpp
@@ -1,16 +1,16 @@
-inline ll LCM(ll a, ll b) {
+inline int LCM(int a, int b) {
 	return a / __gcd(a, b) * b;
 }
 
-inline ll normalize(ll x, ll mod) {
+inline int normalize(int x, int mod) {
 	x %= mod;
 	if (x < 0) x += mod;
 	return x;
 }
 
-struct gcd_type { ll x, y, d; };
+struct gcd_type { int x, y, d; };
 
-gcd_type ex_gcd(ll a, ll b) {
+gcd_type ex_gcd(int a, int b) {
 	if (b == 0) return {1, 0, a};
 	gcd_type pom = ex_gcd(b, a % b);
 	return {pom.y, pom.x - a / b * pom.y, pom.d};
@@ -25,8 +25,8 @@ pair<ll, ll> crt(const vector<int> &vals, const vector<int> &mods) {
 	for(int i = 0; i < vals.size(); i++) {
 		normalize(vals[i], mods[i]);
 	}
-	ll ans = vals[0];
-	ll Lcm = mods[0];
+ int ans = vals[0];
+ int Lcm = mods[0];
 	for(int i = 1; i < vals.size(); i++) {
 		auto pom = ex_gcd(Lcm, mods[i]);
 		int x1 = pom.x;
