+++ maxflow/hungarian.cpp
@@ -4,26 +4,26 @@
 // #jobs must be <= #workers
 // Default finds min cost; to find max cost set all costs[i][j] to -costs[i][j]
 ll HungarianMatch(const vector<vector<ll>>& a) {
-	ll n = a.size()-1;
-	ll m = a[0].size()-1;
+ int n = a.size()-1;
+ int m = a[0].size()-1;
 	vector<ll> u(n+1), v(m+1), p(m+1), way(m+1);
-	for(ll i = 1; i <= n; ++i) {
+	for(int i = 1; i <= n; ++i) {
 		p[0] = i;
-		ll j0 = 0;
+	 int j0 = 0;
 		vector<ll> minv(m+1, inf);
 		vector<char> used(m+1, false);
 		do {
 			used[j0] = true;
-			ll i0 = p[j0], delta = inf, j1;
-			for(ll j = 1; j <= m; ++j)
+		 int i0 = p[j0], delta = inf, j1;
+			for(int j = 1; j <= m; ++j)
 				if(!used[j]) {
-					ll cur = a[i0][j] - u[i0] - v[j];
+				 int cur = a[i0][j] - u[i0] - v[j];
 					if(cur < minv[j])
 						minv[j] = cur, way[j] = j0;
 					if(minv[j] < delta)
 						delta = minv[j], j1 = j;
 				}
-			for(ll j = 0; j <= m; ++j)
+			for(int j = 0; j <= m; ++j)
 				if(used[j])
 					u[p[j]] += delta, v[j] -= delta;
 				else
@@ -31,7 +31,7 @@ ll HungarianMatch(const vector<vector<ll>>& a) {
 			j0 = j1;
 		} while(p[j0] != 0);
 		do {
-			ll j1 = way[j0];
+		 int j1 = way[j0];
 			p[j0] = p[j1];
 			j0 = j1;
 		} while(j0);
@@ -40,7 +40,7 @@ ll HungarianMatch(const vector<vector<ll>>& a) {
 	/*
 	// For each N, it contains the M it selected
 	vector<ll> ans(n+1);
-	for(ll j = 1; j <= m ; ++j)
+	for(int j = 1; j <= m ; ++j)
 		ans[p[j]] = j;
 	*/
