+++ maxflow/minCostMaxFlow.cpp
@@ -1,19 +1,19 @@
 struct mincostmaxflow {
 	struct edge {
-		ll a, b, cap, cost, flow;
+	 int a, b, cap, cost, flow;
 		size_t back;
 	};
 
 	vector<edge> e;
 	vector<vector<ll>> g;
-	ll n, s, t;
-	ll k = inf; // The maximum amount of flow allowed
+ int n, s, t;
+ int k = inf; // The maximum amount of flow allowed
 
 	mincostmaxflow(int _n, int _s, int _t) : n(_n), s(_s), t(_t) {
 		g.resize(n);
 	}
 
-	void addedge(ll a, ll b, ll cap, ll cost) {
+	void addedge(int a, int b, int cap, int cost) {
 		edge e1 = {a,b,cap,cost,0,g[b].size()};
 		edge e2 = {b,a,0,-cost,0,g[a].size()};
 		g[a].push_back((ll) e.size());
@@ -24,18 +24,18 @@ struct mincostmaxflow {
 
 	// Returns {flow,cost}
 	pair<ll,ll> getflow() {
-		ll flow = 0, cost = 0;
+	 int flow = 0, cost = 0;
 		while(flow < k) {
 			vector<ll> id(n, 0);
 			vector<ll> d(n, inf);
 			vector<ll> q(n);
 			vector<ll> p(n);
 			vector<size_t> p_edge(n);
-			ll qh=0, qt=0;
+		 int qh=0, qt=0;
 			q[qt++] = s;
 			d[s] = 0;
 			while(qh != qt) {
-				ll v = q[qh++];
+			 int v = q[qh++];
 				id[v] = 2;
 				if(qh == n) qh = 0;
 				for(size_t i=0; i<g[v].size(); ++i) {
@@ -57,13 +57,13 @@ struct mincostmaxflow {
 				}
 			}
 			if(d[t] == inf) break;
-			ll addflow = k - flow;
-			for(ll v=t; v!=s; v=p[v]) {
-				ll pv = p[v]; size_t pr = p_edge[v];
+		 int addflow = k - flow;
+			for(int v=t; v!=s; v=p[v]) {
+			 int pv = p[v]; size_t pr = p_edge[v];
 				addflow = min(addflow, e[g[pv][pr]].cap - e[g[pv][pr]].flow);
 			}
-			for(ll v=t; v!=s; v=p[v]) {
-				ll pv = p[v]; size_t pr = p_edge[v], r = e[g[pv][pr]].back;
+			for(int v=t; v!=s; v=p[v]) {
+			 int pv = p[v]; size_t pr = p_edge[v], r = e[g[pv][pr]].back;
 				e[g[pv][pr]].flow += addflow;
 				e[g[v][r]].flow -= addflow;
 				cost += e[g[pv][pr]].cost * addflow;
